pyproject.toml
[project]
name = "gemini-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.5",
    "nest-asyncio>=1.6.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
